### 2021-10-05 #3802 @whedon
[PRE REVIEW]: piecewise-regression (aka segmented regression) in Python
**Submitting author:** @chasmani (<a href="http://orcid.org/0000-0002-3800-677X">Charlie Pilgrim</a>)
**Repository:** <a href="https://github.com/chasmani/piecewise-regression" target ="_blank">https://github.com/chasmani/piecewise-regression</a>
**Version:** 1.0.4
