### 2021-10-05 #3802 @kyleniemeyer
@openjournals/dev any ideas what is going on here? I have never seen this behavior
