### 2021-10-05 #3802 @kyleniemeyer
Well, not sure what was going on there at first, but everything seems to be working fine now.

@chasmani we will use this issue to fine an editor and reviewers for your submission. While we are doing that, it looks like your software repository is missing an open source license, which is required for submissions to JOSS. Can you please select an [OSI-approved license](https://opensource.org/licenses) and add it to your repo?
