### 2021-10-26 #3859 @whedon
[REVIEW]: piecewise-regression (aka segmented regression) in Python
**Submitting author:** @chasmani (<a href="http://orcid.org/0000-0002-3800-677X">Charlie Pilgrim</a>)
**Repository:** <a href="https://github.com/chasmani/piecewise-regression" target ="_blank">https://github.com/chasmani/piecewise-regression</a>
**Version:** v1.2.0

@vyasr, @lauravana and @htjb, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
