### 2021-10-31 #3859 @htjb
Dear @chasmani,

Thank you for the interesting submission! The software appears to fill a
gap in the Python library of tools and the paper is a useful contribution to the
literature. I have been able to install and run the code without any issues.

Overall the **repository** is well put together and the code is well written and
documented. I did however have a few comments which I think would make the
code more accessible to the community. I have listed these below since I do not believe they are sufficient to warrant separate *Issues* but I am happy to break them up into *Issues* on the repo if you would prefer. Also happy to discuss any of the points further if clarification is needed.

- The readme is well written however it might benefit from an additional section
detailing the required packages. It would also be good to include a title
and a table at the top of the readme including the current version number and
quick links to the repo and documentation (as is done in [anesthetic](https://github.com/williamjameshandley/anesthetic) and [globalemu](https://github.com/htjb/globalemu)).
I would also refrain from suggesting that people use `sudo pip3 install` in the
installation section since `pip3 install` is typically sufficient. Finally you may
want to add instructions to install the package from source e.g.
```
git clone https://github.com/chasmani/piecewise-regression
cd piecewise_regression
python3 setup.py install --user
```
- Again the code is well written and particularly well documented. However, I think
that it is important that the code meets certain conventions and styles that help
with readability. The code should comply with the [PEP-8](https://www.python.org/dev/peps/pep-0008/) standard. You can check against PEP-8 using [Flake8](https://flake8.pycqa.org/en/latest/) in the terminal (e.g. `flake8 piecewise_regression`).
- There are no clear community guidelines in the repo. You should add a section in
the readme about contributing to the software (again some examples include [anesthetic](https://github.com/williamjameshandley/anesthetic) and [globalemu](https://github.com/htjb/globalemu)).
- The testing suite looks to be well constructed. It is worth adding instructions on
how the user can recover the coverage of the testing suite e.g. by running
```
nosetests --with-coverage --cover-erase --cover-package=piecewise_regression/
```
to the readme. The coverage of the testing suite is a little low at 67% and I would
suggest writing some additional tests to increase the coverage to >90% (higher coverage = higher user confidence). Ideally you should also set up continuous integration as
well which can be done with [github actions](https://docs.github.com/en/actions/automating-builds-and-tests/about-continuous-integration). Although I do not believe this is a specific requirement for JOSS I would advise it. You should refrain from including data directly in the
testing codes as well (e.g. np.array[[100.6466801609725, 100.22604536744389, ...]). It is perfectly reasonable to save this to `.txt` or `.npy` files and upload it to
the repo in a folder `tests/data/` so that it can be read in by the testing scripts.

- The example use case is easy to follow and use. It might be advantageous to put this into a Jupyter notebook and host it on [binder](https://mybinder.org/) for ease of access.

In terms of the **paper**

- More details should be given regarding the example fit. For example reporting and explaining the results of the Davies hypothesis test for this example. It may be advisable to merge the `Example` and `How it works` sections for readability.
- A discussion of the BIC would be nice as well. Perhaps fitting the example with
max_breakpoints=5 and using the BIC to determine which is favourable.
- I would cite the relevant papers for the Davies test.
- If there are any Python implementations of the method using a grid search for the
breakpoints I believe these should be referenced.

I hope you find these comments useful! As I have said overall the code is really nice and I think the additional statistical tests that have been implemented (Davies and BIC) are really useful.

Best,
Harry
