### 2021-11-02 #3859 @galessiorob
Thank you so much for your speedy review @htjb ⚡ I appreciate the thorough documentation on the feedback.

@chasmani, if you can please start by addressing the automated tests and the community guidelines so we can make progress with the actual software. I'll open a PR with some edit suggestions shortly so we can address the paper.

@vyasr please let me know if I can help with anything so you can get started with the review. Appreciate you volunteering your time!
