### 2021-11-03 #3859 @Ebedthan
Dear @chasmani,

Thank you for your submission which describes software that I think will be helpful for the Python community where the lack of such software was a gap.
I have some comments that I think will improve the overall submission. These comments intentionally avoid duplication with comments from @htjb that I hope you will fully take into consideration. I hope you will find it helpful.

First on the repository:
* I would suggest the README start with a headline stating the software name and the aim of the software. You can also add if you want a logo to give a visual identity to your software.
* You can fill the "About" section of your repository to make it easily findable on GitHub and rapidly give a sense of what the repository contains.

Finally on the paper:
* Line 6: add a comma before "or"
* Line 10: change "easy to use" to "easy-to-use"
* line 11: remove the comma before "and hypothesis"
* Line 16-17: typo in "altearnative". I think you wanted "alternative"
* Line 18: for conciseness, I would go with "Many R packages implement" in place of "There are many R packages that implement"
* Line 20: add a comma after "writing"
* Line 20: replace "not" with "no"
* Line 24: replace "in order to" with "to"
* Line 35-36: rephrase the sentence to make it clearer
* Line 38: add a comma after limitation

Thank you for your submission and the software 👍🏾.

Best,
Anicet
