### 2021-11-04 dbc6056
Updated paper and readme following reviewer comments
