### 2021-11-04 #3859 @chasmani
@Ebedthan thank you very much for your review and comments.

**I've made changes based on your comments. Responses are in bold below: **

First on the repository:

-   I would suggest the README start with a headline stating the software name and the aim of the software. You can also add if you want a logo to give a visual identity to your software.
**Title updated. I feel the first paragraph states the aim of the software succinctly. A logo is a nice idea that I will consider adding at a later point**
    You can fill the "About" section of your repository to make it easily findable on GitHub and rapidly give a sense of what the repository contains.
**Thank you that is a good tip. Added an About section and relevant topics**

Finally on the paper:

    Line 6: add a comma before "or"
    Line 10: change "easy to use" to "easy-to-use"
    line 11: remove the comma before "and hypothesis"
    Line 16-17: typo in "altearnative". I think you wanted "alternative"
    Line 18: for conciseness, I would go with "Many R packages implement" in place of "There are many R packages that implement"
    Line 20: add a comma after "writing"
    Line 20: replace "not" with "no"
**not is more correct. There are some resources, just nowhere near as good as in R (and also not worth mentioning as they are e.g. not complete)**
    Line 24: replace "in order to" with "to"
    Line 35-36: rephrase the sentence to make it clearer
    Line 38: add a comma after limitation
**All changed**

**Thanks again. I will address the other reviewer comments today as well.
Charlie**
