### 2021-11-04 33b5104
PEP8 compliance changes
