+++ tests/test_fit.py
@@ -1,41 +1,42 @@
-import copy
+from piecewise_regression.main import Fit, Muggeo
 
-import pandas as pd
 import numpy as np
 import unittest
 
-import os, sys
+import os
+import sys
 sys.path.insert(1, os.path.join(sys.path[0], '..'))
-
-from piecewise_regression.main import Fit, Muggeo
 
 
 class TestFit(unittest.TestCase):
 
     def test_it_works_with_differnet_options(self):
 
-        xx = np.linspace(0,1,100)
-        yy = np.linspace(0,1,100)
+        xx = np.linspace(0, 1, 100)
+        yy = np.linspace(0, 1, 100)
 
-        fit = Fit(xx, yy, n_breakpoints=2, verbose=False, n_boot=3)
-        fit = Fit(xx, yy, n_breakpoints=1, verbose=False, n_boot=3)
-        fit = Fit(xx, yy, n_breakpoints=4, verbose=False, n_boot=3)
-        fit = Fit(xx, yy, n_breakpoints=7, verbose=False, n_boot=3)
-        fit = Fit(xx, yy, start_values=[0.1,0.6], verbose=False)
-        fit = Fit(xx, yy, start_values=[2,3], verbose=False)
-        fit = Fit(xx, yy, start_values=[2], verbose=False)
-
-
+        Fit(xx, yy, n_breakpoints=2, verbose=False, n_boot=3)
+        Fit(xx, yy, n_breakpoints=1, verbose=False, n_boot=3)
+        Fit(xx, yy, n_breakpoints=4, verbose=False, n_boot=3)
+        Fit(xx, yy, n_breakpoints=7, verbose=False, n_boot=3)
+        Fit(xx, yy, start_values=[0.1, 0.6], verbose=False)
+        Fit(xx, yy, start_values=[2, 3], verbose=False)
+        Fit(xx, yy, start_values=[2], verbose=False)
 
     def test_against_muggeo_r_package_data_1(self):
         """
-        Muggeo uses slightly different packages and methods etc, so just check values are very close, not exact
-        Starting from Muggeo's converged breakpoint values, I am iterating once, slight change
-        The NextBreakpoint class is very vanilla, so in this example is getting in some local minima
+        Muggeo uses slightly different packages and methods etc, so just check
+        values are very close, not exact
+        Starting from Muggeo's converged breakpoint values, I am iterating
+        once, slight change
+        The NextBreakpoint class is very vanilla, so in this example is
+        getting in some local minima
         """
 
[…]
