+++ setup.py
@@ -4,9 +4,11 @@ import setuptools
 # The directory containing this file
 HERE = pathlib.Path(__file__).parent
 
+
 def readme():
     with open('docs/README.rst') as f:
         return f.read()
+
 
 # This call to setup() does all the work
 setuptools.setup(
