+++ tests/test_r_squared.py
@@ -1,10 +1,13 @@
 import piecewise_regression.r_squared_calc as r_squared_calc
 import numpy as np
 import unittest
+import imp
 
 import os
 import sys
 sys.path.insert(1, os.path.join(sys.path[0], '..'))
+
+DATA_SOURCE = "tests/data/data.txt"
 
 
 class TestRSquared(unittest.TestCase):
@@ -28,11 +31,10 @@ class TestRSquared(unittest.TestCase):
         r_2_from_sklearn = -6.405023255813953
         self.assertEqual(r_2_from_sklearn, r_2)
 
-        # Test with data from bp fit
-        ff = np.array([100.6466801609725, 100.22604536744389, 99.80541057391528, 99.38477578038668, 98.96414098685806, 98.54350619332945, 98.12287139980084, 97.70223660627224, 97.28160181274363, 96.86096701921502, 96.44033222568642, 96.01969743215781, 95.59906263862919, 95.17842784510059, 94.75779305157198, 94.33715825804337, 93.91652346451477, 93.49588867098616, 93.07525387745756, 92.65461908392894, 92.23398429040033, 91.81334949687172, 91.39271470334312, 90.97207990981451, 90.5514451162859, 90.1308103227573, 89.71017552922868, 89.28954073570007, 88.86890594217147, 88.44827114864286, 88.02763635511425, 87.60700156158565, 87.18636676805704, 86.76573197452842, 86.34509718099982, 85.92446238747121, 85.5038275939426, 85.083192800414, 84.66255800688539, 84.24192321335678, 83.82128841982816, 83.40065362629957, 82.98001883277095, 82.55938403924235, 82.13874924571374, 81.71811445218513, 81.29747965865653, 80.87684486512791, 80.45621007159932, 80.0355752780707, 79.61494048454209, 79.19430569101348, 78.77367089748488, 78.35303610395627, 77.93240131042766, 77.51176651689906, 77.09113172337044, 76.67049692984183, 76.24986213631323, 75.82922734278462, 75.40859254925601, 74.98795775572741, 74.5673229621988, 74.14668816867018, 73.72605337514159, 73.30541858161297, 72.88478378808436, 72.46414899455576, 72.04351420102715, 71.62287940749854, 71.20224461396992, 70.78160982044133, 70.04396951676658, 69.25506553978708, 68.46616156280757, 67.67725758582806, 66.88835360884856, 66.09944963186904, 65.31054565488954, 64.52164167791003, 63.73273770093051, 62.943833723951, 62.15492974697149, 61.366025769991985, 60.577121793012466, 59.78821781603297, 58.99931383905345, 58.210409862073945, 57.42150588509443, 56.63260190811493, 55.84369793113541, 55.054793954155905, 54.2658899771764, 53.47698600019689, 52.688082023217376, 51.899178046237864, 51.11027406925836, 50.321370092278855, 49.532466115299336, 48.74356213831983, 47.95465816134032,
[…]
