+++ tests/test_next_breakpoint.py
@@ -2,10 +2,13 @@ from piecewise_regression.main import NextBreakpoints
 
 import numpy as np
 import unittest
+import imp
 
 import os
 import sys
 sys.path.insert(1, os.path.join(sys.path[0], '..'))
+
+DATA_SOURCE = "tests/data/data.txt"
 
 
 class TestNextBreakpoint(unittest.TestCase):
@@ -20,10 +23,10 @@ class TestNextBreakpoint(unittest.TestCase):
         in some local minima
         """
 
-        xx = np.array([0.0, 0.10050251256281408, 0.20100502512562815, 0.30150753768844224, 0.4020100502512563, 0.5025125628140704, 0.6030150753768845, 0.7035175879396985, 0.8040201005025126, 0.9045226130653267, 1.0050251256281408, 1.105527638190955, 1.206030150753769, 1.306532663316583, 1.407035175879397, 1.5075376884422111, 1.6080402010050252, 1.7085427135678393, 1.8090452261306533, 1.9095477386934674, 2.0100502512562817, 2.1105527638190957, 2.21105527638191, 2.311557788944724, 2.412060301507538, 2.512562814070352, 2.613065326633166, 2.71356783919598, 2.814070351758794, 2.9145728643216082, 3.0150753768844223, 3.1155778894472363, 3.2160804020100504, 3.3165829145728645, 3.4170854271356785, 3.5175879396984926, 3.6180904522613067, 3.7185929648241207, 3.819095477386935, 3.919597989949749, 4.020100502512563, 4.120603015075377, 4.2211055276381915, 4.3216080402010055, 4.42211055276382, 4.522613065326634, 4.623115577889448, 4.723618090452262, 4.824120603015076, 4.92462311557789, 5.025125628140704, 5.125628140703518, 5.226130653266332, 5.326633165829146, 5.42713567839196, 5.527638190954774, 5.628140703517588, 5.728643216080402, 5.8291457286432165, 5.9296482412060305, 6.030150753768845, 6.130653266331659, 6.231155778894473, 6.331658291457287, 6.432160804020101, 6.532663316582915, 6.633165829145729, 6.733668341708543, 6.834170854271357, 6.934673366834171, 7.035175879396985, 7.135678391959799, 7.236180904522613, 7.336683417085427, 7.437185929648241, 7.5376884422110555, 7.63819095477387, 7.738693467336684, 7.839195979899498, 7.939698492462312, 8.040201005025127, 8.14070351758794, 8.241206030150755, 8.341708542713569, 8.442211055276383, 8.542713567839197, 8.643216080402011, 8.743718592964825, 8.84422110552764, 8.944723618090453, 9.045226130653267, 9.145728643216081, 9.246231155778895, 9.34673366834171, 9.447236180904524, 9.547738693467338, 9.648241206030152, 9.748743718592966, 9.84924623115578, 9.949748743718594, 10.050251256281408,
[…]
