+++ paper/paper.bib
@@ -39,3 +39,44 @@
   year={2001},
   publisher={Wiley Online Library}
 }
+
+@article{davies1987hypothesis,
+  title={Hypothesis testing when a nuisance parameter is present only under the alternative},
+  author={Davies, Robert B},
+  journal={Biometrika},
+  volume={74},
+  number={1},
+  pages={33--43},
+  year={1987},
+  publisher={Oxford University Press}
+}
+
+@article{wit2012all,
+  title={‘All models are wrong...’: an introduction to model uncertainty},
+  author={Wit, Ernst and Heuvel, Edwin van den and Romeijn, Jan-Willem},
+  journal={Statistica Neerlandica},
+  volume={66},
+  number={3},
+  pages={217--236},
+  year={2012},
+  publisher={Wiley Online Library}
+}
+
+@article{virtanen2020scipy,
+  title={SciPy 1.0: fundamental algorithms for scientific computing in Python},
+  author={Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and others},
+  journal={Nature methods},
+  volume={17},
+  number={3},
+  pages={261--272},
+  year={2020},
+  publisher={Nature Publishing Group}
+}
+
+@article{jekel2019pwlf,
+  title={PWLF: a python library for fitting 1D continuous piecewise linear functions},
+  author={Jekel, Charles F and Venter, Gerhard},
+  journal={URL: https://github. com/cjekel/piecewise\_linear\_fit\_py},
+  year={2019}
+}
+
