+++ piecewise_regression/model_selection.py
@@ -117,5 +117,18 @@ class ModelSelection:
 
 
 if __name__ == "__main__":
+    np.random.seed(0)
 
-    pass
+    alpha = 10
+    intercept = 100
+
+    n_points = 200
+
+    xx = np.linspace(0, 20, n_points)
+
+    yy = intercept + alpha*xx
+    yy += np.random.normal(size=n_points)
+
+    
+
+    ModelSelection(xx, yy, max_breakpoints=6)
