+++ paper/paper.bib
@@ -62,21 +62,9 @@
   publisher={Wiley Online Library}
 }
 
-@article{virtanen2020scipy,
-  title={SciPy 1.0: fundamental algorithms for scientific computing in Python},
-  author={Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and others},
-  journal={Nature methods},
-  volume={17},
-  number={3},
-  pages={261--272},
-  year={2020},
-  publisher={Nature Publishing Group}
-}
-
 @article{jekel2019pwlf,
   title={PWLF: a python library for fitting 1D continuous piecewise linear functions},
   author={Jekel, Charles F and Venter, Gerhard},
   journal={URL: https://github. com/cjekel/piecewise\_linear\_fit\_py},
   year={2019}
 }
-
