+++ requirements.txt
@@ -1,5 +1,5 @@
 matplotlib==3.3.4
 nose==1.3.7
 numpy==1.19.5
-scipy==1.4.1
+scipy==1.6.1
 statsmodels==0.11.1
