### 2021-11-08 220e7cb
Updtaed statsmodels requirement
