+++ requirements.txt
@@ -2,4 +2,4 @@ matplotlib==3.3.4
 nose==1.3.7
 numpy==1.19.5
 scipy==1.6.1
-statsmodels==0.11.1
+statsmodels==0.13.0
