+++ tests/test_data_validation.py
@@ -108,6 +108,9 @@ class TestDataValidation(unittest.TestCase):
                 validate_list_of_numbers,
                 **test_kwargs)
 
+    def test_fail(self):
+        self.assertEqual(1,0)
+
 
 if __name__ == '__main__':
     unittest.main()
