+++ docs/README.rst
@@ -7,6 +7,9 @@ piecewise-regression (aka segmented regression) in python
 :Github: https://github.com/chasmani/piecewise-regression
 :Documentation: https://readthedocs.org/projects/piecewise-regression/
 
+.. image:: https://github.com/williamjameshandley/anesthetic/workflows/CI/badge.svg?branch=master
+   :target: https://github.com/williamjameshandley/anesthetic/actions?query=workflow%3ACI+branch%3Amaster
+   :alt: Build Status
 .. image:: https://codecov.io/gh/chasmani/piecewise-regression/branch/master/graph/badge.svg
    :target: https://codecov.io/gh/chasmani/piecewise-regression
    :alt: Test Coverage Status
