### 2021-11-15 859ee08
Added CI to workflow
