+++ travis.yml
@@ -1,0 +1,11 @@
+language: python
+python:
+  - "3.7"
+  - "3.8"
+  - "3.9"
+# command to install dependencies
+install:
+  - pip install -r requirements.txt
+# command to run tests
+script:
+  - pytest
