### 2021-11-15 10ce3c4
Added nosetests to travis
