+++ travis.yml
@@ -8,4 +8,4 @@ install:
   - pip install -r requirements.txt
 # command to run tests
 script:
-  - pytest
+  - nosetests --with-coverage
