### 2021-11-15 05eae8b
Uddated CI options
