### 2021-11-15 33c7cac
Updated os options
