+++ .github/workflows/codecov.yml
@@ -26,9 +26,8 @@ jobs:
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v2
       with:
-        directory: ./coverage/reports/
         env_vars: OS,PYTHON
         fail_ci_if_error: true
         name: codecov-umbrella
-        path_to_write_report: ./coverage/codecov_report.txt
+        path_to_write_report: tests/coverage/codecov_report.txt
         verbose: true
