+++ .github/workflows/codecov.yml
@@ -19,15 +19,18 @@ jobs:
     - name: Generate coverage report
       run: |
         python -m pip install --upgrade pip
-        python -m pip install flake8 pytest
+        python -m pip install flake8 pytest pytest-cov
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
-        pip install pytest-cov
         pytest --cov=./ --cov-report=xml
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v2
       with:
         env_vars: OS,PYTHON
+        directory: ./coverage/reports/
+        env_vars: OS,PYTHON
         fail_ci_if_error: true
+        files: ./coverage1.xml,./coverage2.xml
+        flags: unittests
         name: codecov-umbrella
-        path_to_write_report: tests/coverage/codecov_report.txt
+        path_to_write_report: ./coverage/codecov_report.txt
         verbose: true
