+++ .github/workflows/codecov.yml
@@ -27,7 +27,6 @@ jobs:
       with:
         env_vars: OS,PYTHON
         directory: ./coverage/reports/
-        env_vars: OS,PYTHON
         fail_ci_if_error: true
         files: ./coverage1.xml,./coverage2.xml
         flags: unittests
