+++ .github/workflows/codecov.yml
@@ -26,10 +26,7 @@ jobs:
       uses: codecov/codecov-action@v2
       with:
         env_vars: OS,PYTHON
-        directory: ./coverage/reports/
         fail_ci_if_error: true
-        files: ./coverage1.xml,./coverage2.xml
-        flags: unittests
         name: codecov-umbrella
-        path_to_write_report: ./coverage/codecov_report.txt
+        path_to_write_report: tests/coverage/codecov_report.txt
         verbose: true
