+++ paper/paper.md
@@ -47,7 +47,7 @@ where given some data, $x$, $y$, we are trying to estimate the gradient of the f
 \end{equation}
 
 
-This is now a linear relationship and we can find a new breakpoint estimate, $\psi^{(1)}$, through linear regression. We iterate in this way until the breakpoint estimate converges, at which point we stop the algorithm. \autoref{eqn:expansion} is the linear approximation for one breakpoint. If considering multiple breakpoints, the same derivation is followed but instead using a multivariate Taylor expansion around an initial guess for each of the breakpoints. 
+This is now a linear relationship and we can find a new breakpoint estimate, $\psi^{(1)}$, through linear regression. We iterate in this way until the breakpoint estimate converges, at which point we stop the algorithm. \autoref{eqn:expansion} is the linear approximation for one breakpoint. If considering multiple breakpoints, the same approach is followed but instead using a multivariate Taylor expansion around an initial guess for each of the breakpoints. 
 
 Muggeo's iterative algorithm is not guaranteed to converge on a globally optimal solution. Instead, it can converge to local optima or diverge. To address this limitation, we also implement bootstrap restarting [@wood2001minimizing], again following Muggeo's approach [@muggeo2008segmented]. The bootstrap restarting algorithm generates a non-parametric bootstrap of the data through resampling, which is then used to find new breakpoint values that may find a better global solution. This is repeated several times to escape local optima.  
 
