### 2021-11-15 #3859 @chasmani
Thank you @htjb for your review and comments. There were some good tips to improve the codebase. (I will address @vyasr 's comments separately).

I've addressed those and included responses in bold below. I've implemented the majority of the suggestions.

1. The readme is well written however it might benefit from an additional section detailing the required packages. It would also be good to include a title and a table at the top of the readme including the current version number and quick links to the repo and documentation (as is done in anesthetic and globalemu).
**I added this. Also added some of the badges.**

2. I would also refrain from suggesting that people use sudo pip3 install in the installation section since pip3 install is typically sufficient.
**Done**

3. Finally you may want to add instructions to install the package from source e.g.
    git clone https://github.com/chasmani/piecewise-regression
    cd piecewise_regression
    python3 setup.py install --user
**I added this to the bottom of the readme - I don't want to add too much to the top as it may be confusing or unclear to some users**

4. Again the code is well written and particularly well documented. However, I think that it is important that the code meets certain conventions and styles that help with readability. The code should comply with the PEP-8 standard. You can check against PEP-8 using Flake8 in the terminal (e.g. flake8 piecewise_regression).
**Updated to PEP8 compliance in the vast majority of instances**
**A small number of lines do not comply because it would make the code less readable e.g. using descriptive but long variable names means some lines og over the recommend length, but changing those variables to be shorter would make the code less readable**

5. There are no clear community guidelines in the repo. You should add a section in the readme about contributing to the software (again some examples include anesthetic and globalemu).
**Done**

6. The testing suite looks to be well constructed. It is worth adding instructions on how the user can recover the coverage of the testing suite e.g. by running

    nosetests --with-coverage --cover-erase --cover-package=piecewise_regression/
to the readme.
**Done **

7. The coverage of the testing suite is a little low at 67% and I would suggest writing some additional tests to increase the coverage to >90% (higher coverage = higher user confidence).
**Done. Increased to 97%**

8. Ideally you should also set up continuous integration as well which can be done with github actions. Although I do not believe this is a specific requirement for JOSS I would advise it.
**Thanks for the suggestion. I've not used Continuous Integration before and it looks great. I implemented CI to automiatcially run tests upon git push, tests for python 3.7,3.8,3.9. Currently passing. That is very useful for automatically checking that the package works across python versions. Also it automatically uploads the text coverage to codecov.**

9. You should refrain from including data directly in the
testing codes as well (e.g. np.array[[100.6466801609725, 100.22604536744389, ...]). It is perfectly reasonable to save this to .txt or .npy files and upload it to
the repo in a folder tests/data/ so that it can be read in by the testing scripts.
**Done**

10. The example use case is easy to follow and use. It might be advantageous to put this into a Jupyter notebook and host it on binder for ease of access.
**I don't think this is necessary, considering that the package is quite simple to use. It is more overhead for things to update following changes. And duplicating the code increases the chance that the code example in the readme and the Jupyter notebook will diverge at some point in the future, which could be confusing for users. I prefer to have one single instance of the code example, following DRY. I could generate the README code and the Jupyter notebook from a single source but that feels like over engineering. I can add this if it is a deal breaker.**

In terms of the paper

1. More details should be given regarding the example fit. For example reporting and explaining the results of the Davies hypothesis test for this example.
**Done.**

2. It may be advisable to merge the Example and How it works sections for readability.
**I'd prefer to keep these seperate because I anticipate that some users will be interested in the underlying Maths and others will be put off by it. As it is, the users who are not mathematically inclined can clearly see that the How It Works section is not crucial and they can skip it.**

3. A discussion of the BIC would be nice as well. Perhaps fitting the example with max_breakpoints=5 and using the BIC to determine which is favourable.
**Included. I breifly explained the rationale of using the BIC, and reported the result of the preferred number of breakpoints using the same data as shown in the exmaple figure.**

4. I would cite the relevant papers for the Davies test.
**Yes good point. Done.**

5. If there are any Python implementations of the method using a grid search for the breakpoints I believe these should be referenced.
**Done**
