### 2021-11-15 4fb1d90
Converted nosetests to pytest in docs
