+++ paper/paper.bib
@@ -19,6 +19,42 @@
   year={2008}
 }
 
+@article{toms2003piecewise,
+  title={Piecewise regression: a tool for identifying ecological thresholds},
+  author={Toms, Judith D and Lesperance, Mary L},
+  journal={Ecology},
+  volume={84},
+  number={8},
+  pages={2034--2041},
+  year={2003},
+  publisher={Wiley Online Library}
+}
+
+@article{ryan2002defining,
+  title={Defining phases of bedload transport using piecewise regression},
+  author={Ryan, Sandra E and Porth, Laurie S and Troendle, CA},
+  journal={Earth Surface Processes and Landforms},
+  volume={27},
+  number={9},
+  pages={971--990},
+  year={2002},
+  publisher={Wiley Online Library}
+}
+
+
+
+@article{wagner2002segmented,
+  title={Segmented regression analysis of interrupted time series studies in medication use research},
+  author={Wagner, Anita K and Soumerai, Stephen B and Zhang, Fang and Ross-Degnan, Dennis},
+  journal={Journal of clinical pharmacy and therapeutics},
+  volume={27},
+  number={4},
+  pages={299--309},
+  year={2002},
+  publisher={Wiley Online Library}
+}
+
+
 @inproceedings{seabold2010statsmodels,
   title={Statsmodels: Econometric and statistical modeling with python},
   author={Seabold, Skipper and Perktold, Josef},
