+++ paper/paper.bib
@@ -41,6 +41,32 @@
   publisher={Wiley Online Library}
 }
 
+@article{virtanen2020scipy,
+  title={SciPy 1.0: fundamental algorithms for scientific computing in Python},
+  author={Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and others},
+  journal={Nature methods},
+  volume={17},
+  number={3},
+  pages={261--272},
+  year={2020},
+  publisher={Nature Publishing Group}
+}
+
+@article{newville2016lmfit,
+  title={LMFIT: Non-linear least-square minimization and curve-fitting for Python},
+  author={Newville, Matthew and Stensitzki, Till and Allen, Daniel B and Rawlik, Michal and Ingargiola, Antonino and Nelson, Andrew},
+  journal={Astrophysics Source Code Library},
+  pages={ascl--1606},
+  year={2016}
+}
+
+@article{jekel2019pwlf,
+  title={PWLF: a python library for fitting 1D continuous piecewise linear functions},
+  author={Jekel, Charles F and Venter, Gerhard},
+  journal={URL: https://github. com/cjekel/piecewise\_linear\_fit\_py},
+  year={2019}
+}
+
 
 
 @article{wagner2002segmented,
