+++ docs/README.rst
@@ -26,6 +26,7 @@ piecewise-regression (aka segmented regression) in python
    :target: https://github.com/chasmani/piecewise-regresssion/blob/master/LICENSE
    :alt: License information
 
+
 Easy-to-use piecewise regression (aka segmented regression) in Python. For fitting straight lines to data where there is one or more changes in gradient (known as breakpoints). Based on Muggeo "Estimating regression models with unknown break-points" (2003). 
 
 For example:
