+++ tests/test_next_breakpoint.py
@@ -2,7 +2,7 @@ from piecewise_regression.main import NextBreakpoints
 
 import numpy as np
 import unittest
-import imp
+from importlib.machinery import SourceFileLoader
 
 import os
 import sys
@@ -23,7 +23,7 @@ class TestNextBreakpoint(unittest.TestCase):
         in some local minima
         """
 
-        data = imp.load_source('data', DATA_SOURCE)
+        data = SourceFileLoader('data', DATA_SOURCE).load_module()
 
         xx = np.array(data.MUGGEO_1_XX)
         yy = np.array(data.MUGGEO_1_YY)
@@ -115,7 +115,7 @@ class TestNextBreakpoint(unittest.TestCase):
         getting in some local minima
         """
 
-        data = imp.load_source('data', DATA_SOURCE)
+        data = SourceFileLoader('data', DATA_SOURCE).load_module()
 
         xx = np.array(data.MUGGEO_2_XX)
         yy = np.array(data.MUGGEO_2_YY)
