+++ tests/test_fit.py
@@ -3,7 +3,7 @@ from piecewise_regression.main import Fit, Muggeo
 import numpy as np
 import unittest
 import matplotlib.pyplot as plt
-import imp
+from importlib.machinery import SourceFileLoader
 
 import os
 import sys
@@ -37,7 +37,7 @@ class TestFit(unittest.TestCase):
         getting in some local minima
         """
 
-        data = imp.load_source('data', DATA_SOURCE)
+        data = SourceFileLoader('data', DATA_SOURCE).load_module()
 
         xx = np.array(data.MUGGEO_1_XX)
         yy = np.array(data.MUGGEO_1_YY)
@@ -169,7 +169,7 @@ class TestFit(unittest.TestCase):
         getting in some local minima
         """
 
-        data = imp.load_source('data', DATA_SOURCE)
+        data = SourceFileLoader('data', DATA_SOURCE).load_module()
 
         xx = np.array(data.MUGGEO_2_XX)
         yy = np.array(data.MUGGEO_2_YY)
@@ -246,7 +246,7 @@ class TestFit(unittest.TestCase):
         getting in some local minima
         """
 
-        data = imp.load_source('data', DATA_SOURCE)
+        data = SourceFileLoader('data', DATA_SOURCE).load_module()
 
         xx = np.array(data.MUGGEO_2_XX)
         yy = np.array(data.MUGGEO_2_YY)
@@ -316,7 +316,7 @@ class TestFit(unittest.TestCase):
         """
         np.random.seed(2)
 
-        data = imp.load_source('data', DATA_SOURCE)
+        data = SourceFileLoader('data', DATA_SOURCE).load_module()
 
         xx = np.array(data.MUGGEO_1_XX)
         yy = np.array(data.MUGGEO_1_YY)
@@ -338,7 +338,7 @@ class TestFit(unittest.TestCase):
 
         np.random.seed(2)
 
-        data = imp.load_source('data', DATA_SOURCE)
+        data = SourceFileLoader('data', DATA_SOURCE).load_module()
 
         xx = np.array(data.MUGGEO_1_XX)
         yy = np.array(data.MUGGEO_1_YY)
@@ -403,7 +403,7 @@ class TestPlots(unittest.TestCase):
 
         np.random.seed(2)
 
-        data = imp.load_source('data', DATA_SOURCE)
+        data = SourceFileLoader('data', DATA_SOURCE).load_module()
 
         xx = np.array(data.MUGGEO_1_XX)
         yy = np.array(data.MUGGEO_1_YY)
@@ -428,7 +428,7 @@ class TestPlots(unittest.TestCase):
 
         np.random.seed(2)
 
-        data = imp.load_source('data', DATA_SOURCE)
+        data = SourceFileLoader('data', DATA_SOURCE).load_module()
 
         xx = np.array(data.MUGGEO_1_XX)
         yy = np.array(data.MUGGEO_1_YY)
@@ -455,7 +455,7 @@ class TestPlots(unittest.TestCase):
 
         np.random.seed(2)
 
-        data = imp.load_source('data', DATA_SOURCE)
+        data = SourceFileLoader('data', DATA_SOURCE).load_module()
 
         xx = np.array(data.MUGGEO_1_XX)
         yy = np.array(data.MUGGEO_1_YY)
@@ -478,7 +478,7 @@ class TestPlots(unittest.TestCase):
 
         np.random.seed(2)
 
-        data = imp.load_source('data', DATA_SOURCE)
+        data = SourceFileLoader('data', DATA_SOURCE).load_module()
 
         xx = np.array(data.MUGGEO_1_XX)
         yy = np.array(data.MUGGEO_1_YY)
