+++ paper/Vyas review reply.txt
@@ -14,17 +14,15 @@ Requirements:
 3. I mentioned pwlf at the top of this review, I think you should at least mention its existence since it is closely related and is a Python option unlike the R packages you currently mention. It might also be nice to include a rough comparison of the different methodologies. IIRC that package uses Lagrange multipliers for local solutions, combined with a differential evolution algorithm for global optimization, but I haven't looked at it in a long time. Perhaps there are others as well, and I think the paper would benefit from a survey to see if such packages exist.
 **Done, good call.**
 
-
 4a. The first sentence of "How it works" seems out of place. 
 **Done**
 
 4b. Stating that we don't need to know the theory seems like a better fit for the summary by stating something like "this package hides the mathematical complexities of constructing a piecewise fit behind a simple interface" or so.
-**I;ve removed the lone from "How It Works". I decided not to add anythig to the summary for conciseness, I think "easy-to-use" covers it**
-
+**I've removed the line from "How It Works". I decided not to add anythig to the summary for conciseness, I think "easy-to-use" covers it**
 
 5. The example presents a raw plot with no context. Please give some description of the generating model, hopefully in the context of some realistic process that it may represent.
 **Agreed. I explained the data generation process as a pieceiwise linear model with Gaussian noise. I gave the example of observations of a physical process with noise due to sampling errors, so the noise could reasonably be expected to be Gaussian and constant.**
-**That is quite an abstract example I know. I think that is best because it represents the general nature of the fitting algorithm. If I generate data based on a very specific real process then that could require specific caveats that would detract from the general nature of the algorithm, and I do not have deep domain expertise in a suitable area to talk authoritatively about the forms of data.**
+**That is quite an abstract example, with the benefit that it represents the general nature of the fitting algorithm. I thought about generating data based on a more specific real process, but that could require specific caveats that would detract from the general nature of the algorithm, and I do not have deep domain expertise in a suitable area to talk authoritatively about the forms of data.**
 
 6. We need to see a few more references for the Davies and BIC tests.
 **Done**
@@ -36,7 +34,6 @@ Nice to have:
 
 8. Have you considered supporting conda-forge? For a pure Python package like yours it should be relatively trivial, you just need to create a feedstock. Not critical, just something to consider since it's a popular installation method these days.
 **TO DO ** 
-
 
 9. While your tests run fine, I would recommend switching away from nose to either raw unittest. nose has not been supported for >5 years last I checked. pytest would be a more modern alternative, but it would require more effort to restructure your tests.
 **Done. Now uses pytest (tis was also easier to integrate with codecov.io)**
@@ -54,5 +51,3 @@ Nice to have:
 
 13. You could embed your examples along with the code to generate them on your ReadTheDocs. If you're feeling adventurous, nbsphinx would let you create a notebook with the example and embed that directly.
 **Done**
-
-
