+++ docs/README.rst
@@ -39,13 +39,19 @@ There are some code examples below, and more in this |colab_link|.
 
 .. |colab_link| raw:: html
 
-   <a href="https://colab.research.google.com/drive/1Pwv6LqwZU8Zbl0VZH6cwOTwoRzm3CPPC#offline=true&sandboxMode=true" target="_blank">Google Colab Jupyter Notebook</a>ß
+   <a href="https://colab.research.google.com/drive/1Pwv6LqwZU8Zbl0VZH6cwOTwoRzm3CPPC#offline=true&sandboxMode=true" target="_blank">Google Colab Jupyter Notebook</a>
 
 
 Installation
 ========================
 
-You can install piecewise-regression using python's `pip package index <https://pypi.org/project/piecewise-regression/>`_
+You can install piecewise-regression using python's |pip_link|
+
+.. |pip_link| raw:: html
+
+   <a href="https://pypi.org/project/piecewise-regression" target="_blank">pip package index</a>
+
+::
 
     pip install piecewise-regression
 
