+++ setup.py
@@ -13,7 +13,7 @@ def readme():
 # This call to setup() does all the work
 setuptools.setup(
     name="piecewise-regression",
-    version="1.0.4",
+    version="1.0.5",
     description="piecewise (segmented) regression in python",
     long_description=readme(),
     long_description_content_type="text/markdown",
@@ -25,6 +25,8 @@ setuptools.setup(
         "License :: OSI Approved :: MIT License",
         "Programming Language :: Python :: 3",
         "Programming Language :: Python :: 3.7",
+        "Programming Language :: Python :: 3.8",
+        "Programming Language :: Python :: 3.9",
     ],
     packages=setuptools.find_packages(),
     install_requires=["numpy", "matplotlib", "scipy", "statsmodels"],
