### 2021-11-17 d6cbec5
Minimised the pypi docs because images were not showing properly
