### 2021-11-17 ae2b51d
Using a simple link in the pip docs
