### 2021-11-18 30f4f21
Added 50% chance of starting from new values in bootstrap restarting algorithm, and incresed n_boot. Improves chance of dingin a glpbal optima
