### 2021-11-18 89b1174
Updated modelselection to save both the model data and the actal model Fit objects themselves
