+++ docs/README.rst
@@ -144,7 +144,7 @@ The package implements Muggeo's iterative algorithm (Muggeo "Estimating regressi
 
 Muggeo's method doesn't always converge on the best solution - sometimes it finds a locally optimal solution or doesn't converge at all. For this reason the Fit method also implements a process called bootstrap restarting. This involves taking a bootstrap resample of the data, then using this bootstrapped data to try and find a better solution. The number of times this runs can be controlled with `n_boot`. To run the Fit without bootstrap restarting, set `n_boot=0`.  
 
-If you don't have good guesses for inital breakpoints, you can just set the number of e.g. `n_breakpoints=3`. in this case the algorithm will randomly generate start_values for breakpoints until it finds a solution that converges (up to `n_boot` times). This is a good option if the algorithm is otherwise not converging. Be aware that the start_values can influence the final converged model, so setting them randomly in this way may give different results on different runs, epecially if the breakpoint positions are not very clear from the data. 
+If you don't have good guesses for inital breakpoints, you can just set the number of e.g. `n_breakpoints=3`. In this case the algorithm will randomly generate start_values. The bootstrap restarting algorithm will now, with a 50% chance, either use the best currently converged breakpoints or randomly generate new start_values. In this way local optima are escaped in two ways and the algorithm will usually converge to a global optima. Even with these measures the algorithm is not completely guranteed to converge to global optima, as is often the case with fitting non-linear models. Increasing 'n_boot' increases the probability of global convergence, at the cost of computation time.
 
 Model Selection
 ==========================
