### 2021-11-18 f8b2528
Updated summary fucntion to just print
