### 2021-11-18 0d5577a
Reverted to return summary as well to pass tests
