+++ piecewise_regression/main.py
@@ -1011,6 +1011,8 @@ class Fit:
                 run_count += 1
             print(summary)
 
+            return summary
+
         else:
             header = "\n{:^70}\n".format("Breakpoint Regression Results")
 
@@ -1109,6 +1111,8 @@ class Fit:
             summary = header + overview + table + davies_result + "\n\n"
 
             print(summary)
+
+            return summary
 
 
 if __name__ == "__main__":
