### 2021-11-18 4439081
Simplified setuppy and docs
