+++ setup.py
@@ -1,21 +1,17 @@
-import pathlib
 import setuptools
-
-# The directory containing this file
-HERE = pathlib.Path(__file__).parent
-
-
-def readme():
-    with open('docs/PIP_DOC.rst') as f:
-        return f.read()
-
 
 # This call to setup() does all the work
 setuptools.setup(
     name="piecewise-regression",
-    version="1.1.0",
+    version="1.1.1",
     description="piecewise (segmented) regression in python",
-    long_description=readme(),
+    long_description="""
+        piecewise-regression provides tools for fitting continuous straight line \
+        models to data with breakpoint(s) where the gradient changes. 
+
+        For docs and more information, visit the Github repo at \
+        https://github.com/chasmani/piecewise-regression.
+        """,
     long_description_content_type="text/markdown",
     url="https://github.com/chasmani/piecewise-regression",
     author="Charlie Pilgrim",
@@ -31,3 +27,4 @@ setuptools.setup(
     packages=setuptools.find_packages(),
     install_requires=["numpy", "matplotlib", "scipy", "statsmodels"],
 )
+
