+++ docs/PIP_DOC.rst
@@ -1,6 +1,0 @@
-
-
-piecewise-regression provides tools for fitting continuous straight line models to data with breakpoint(s) where the gradient changes. 
-
-For docs and more information, visit the Github repo at https://github.com/chasmani/piecewise-regression.
-
