+++ meta.yaml
@@ -1,0 +1,46 @@
+{% set name = "piecewise-regression" %}
+{% set version = "1.0.8" %}
+
+
+package:
+  name: {{ name|lower }}
+  version: {{ version }}
+
+source:
+  url: https://pypi.io/packages/source/{{ name[0] }}/{{ name }}/piecewise-regression-{{ version }}.tar.gz
+  sha256: 73541954c71253cbe6bdc0b768a27408df649bd6786ee6676cb30fe1290085b6
+
+build:
+  number: 0
+  noarch: python
+  script: {{ PYTHON }} -m pip install . -vv
+
+requirements:
+  host:
+    - pip
+    - python >=3.6
+  run:
+    - matplotlib-base
+    - numpy
+    - python >=3.6
+    - scipy
+    - statsmodels
+
+test:
+  imports:
+    - piecewise_regression
+    - tests
+  commands:
+    - pip check
+  requires:
+    - pip
+
+about:
+  home: https://github.com/chasmani/piecewise-regression
+  summary: piecewise (segmented) regression in python
+  license: MIT
+  license_file: LICENSE.txt
+
+extra:
+  recipe-maintainers:
+    - chasmani
