### 2021-11-18 95f65d4
Small changes to paper and added afiliation
