+++ paper/paper.md
@@ -9,11 +9,13 @@ tags:
 authors:
   - name: Charlie Pilgrim
     orcid: 0000-0002-3800-677X
-    affiliation: "1" 
+    affiliation: "1, 2" 
 affiliations:
- - name: Centre for Doctoral Training in Mathematics for Real-World Systems, University of Warwick 
+ - name: Centre for Doctoral Training in Mathematics for Real-World Systems, University of Warwick, Coventry, UK
    index: 1
-date: 4 October 2021
+ - name: The Alan Turing Institute, London, UK
+   index: 2
+date: 18 November 2021
 bibliography: paper.bib
 
 ---
@@ -28,7 +30,7 @@ A common problem in many fields is to fit a continuous straight line model to da
 
 # Example
 
-An example plot is shown in \autoref{fig:example}. Data was generated with 3 breakpoints and some noise, and a model was then fit to that data with 3 breakpoints. The plot shows the maximum likelihood estimators for the straight line segments and breakpoint positions. The package automatically carries out a Davies hypothesis test [@davies1987hypothesis] for the existence of at least 1 breakpoint, in this example finding strong evidence for breakpoints with $p<0.001$.
+An example plot is shown in \autoref{fig:example}. Data was generated with 3 breakpoints and some noise, and a model was then fit to that data. The plot shows the maximum likelihood estimators for the straight line segments and breakpoint positions. The package automatically carries out a Davies hypothesis test [@davies1987hypothesis] for the existence of at least 1 breakpoint, in this example finding strong evidence for breakpoints with $p<0.001$.
 
 ![An example model fit (red line) to data (grey markers). The estimated breakpoint positions (blue lines) and confidence intervals (shaded blue regions) are shown. The data was generated using a piecewise linear model with a constant level of Gaussian noise. For example, this could represent observations with sampling error of some physical process that undergoes phase transitions. \label{fig:example}](example.png)
 
@@ -47,9 +49,9 @@ where given some data, $x$, $y$, we are trying to estimate the gradient of the f
 \end{equation}
 
 
-This is now a linear relationship and we can find a new breakpoint estimate, $\psi^{(1)}$, through ordinary linear regression using statsmodels [@seabold2010statsmodels]. We iterate in this way until the breakpoint estimate converges, at which point we stop the algorithm. \autoref{eqn:expansion} is the linear approximation for one breakpoint. If considering multiple breakpoints, the same approach is followed but instead using a multivariate Taylor expansion around an initial guess for each of the breakpoints. 
+This is now a linear relationship and we can find a new breakpoint estimate, $\psi^{(1)}$, through ordinary linear regression using the statsmodels python package [@seabold2010statsmodels]. We iterate in this way until the breakpoint estimate converges, at which point we stop the algorithm. If considering multiple breakpoints, the same approach is followed but instead using a multivariate Taylor expansion around an initial guess for each of the breakpoints. 
 
-Muggeo's iterative algorithm is not guaranteed to converge on a globally optimal solution. Instead, it can converge to local optima or diverge. To address this limitation, we also implement bootstrap restarting [@wood2001minimizing], again following Muggeo's approach [@muggeo2008segmented]. The bootstrap restarting algorithm generates a non-parametric bootstrap of the data through resampling, which is then used to find new breakpoint values that may find a better global solution. This is repeated several times to escape local optima.
[…]
