### 2021-11-19 #3859 @chasmani
Thank you @vyasr @Ebedthan @htjb for your reviews, and everyone else for managing the submission.

I really think that your comments have significantly improved the paper and the code, and especially the workflows around making code changes.

This is my first paper at JOSS, I'm not sure what the next step is.
