### 2021-11-22 #3859 @htjb
Dear @chasmani,

Thank you for the response to my comments and apologies I haven't had a chance to look at this work again sooner.

It looks like you have made some good progress with the repository and the testing/coverage looks great. The changes that have been made to the paper are also really helpful. Particularly the more detailed statement of need is great and the google colab notebook will help people get to grips with the software.

I am happy with the state of the repo and paper and have ticked my checklist off. I am still a bit concerned about the PEP8 compatibility however the code is definitely readable and I wouldn't want this to delay acceptance. I will leave this up to @galessiorob but have detailed some specifics below.

When I run flake8 on the updated code I still see a number of inconsistencies with the standard. An example from `data_validation.py` that addresses your concerns about variable names is the following line

``` python
value_error_text = "{} must be a list of numbers with minimum length {}".format(var_name, min_length)
```
which can be re-written to be PEP8 compatible as

``` python
value_error_text = \
  "{} must be a list of numbers with minimum length {}".format(
    var_name, min_length)
```

Similarly there are a lot of imports that don't get used for example in `model_selection.py` and `__init__.py`. `__init__.py` is typically left empty if I recall. You should fix the white space around operators as well and make sure that you leave enough blank lines around functions. Running `flake8 piecewise_regression/ --select=E,F` in the terminal should not return anything. I know this probably sounds very pedantic but not only does it help with readability it also gives a well known standard of code to aim for when people want to contribute via issues and pull requests.

You can also add flake8 to the github actions.

Finally, you should probably remove the file `Vyas review reply.txt`
in `paper/` on the repo.

The package is really helpful and the repo/paper are looking great as I have said.
I will be looking out for opportunities to use this in my work!

Best,
Harry
