+++ docs/README.rst
@@ -28,14 +28,14 @@ piecewise-regression (aka segmented regression) in python
 
 |
 
-Easy-to-use piecewise regression (aka segmented regression) in Python. For fitting straight lines to data where there is one or more changes in gradient (known as breakpoints). Based on Muggeo's paper "Estimating regression models with unknown break-points" (2003). 
+Easy-to-use piecewise regression (aka segmented regression) in Python based on Muggeo's paper "Estimating regression models with unknown break-points" (2003) for fitting straight lines to data where there are one or more gradient changes known as breakpoints.
 
-For example:
+Example:
 
 .. image:: https://raw.githubusercontent.com/chasmani/piecewise-regression/master/paper/example.png
     :alt: basic-example-plot-github
 
-There are some code examples below, and more in this `Google Colab Jupyter Notebook <https://colab.research.google.com/drive/1Pwv6LqwZU8Zbl0VZH6cwOTwoRzm3CPPC#offline=true&sandboxMode=true/>`_.
+Code examples below, and more in this `Google Colab Jupyter Notebook <https://colab.research.google.com/drive/1Pwv6LqwZU8Zbl0VZH6cwOTwoRzm3CPPC#offline=true&sandboxMode=true/>`_.
 
 Installation
 ========================
@@ -49,7 +49,7 @@ The package is tested on Python 3.7, 3.8 and 3.9.
 Getting started
 ========================
 
-The package requires some x and y data to fit. You also need to specify either a) some initial breakpoint guesses as `start_values` or b) how many breakpoints you want to fit as `n_breakpoints` (or both). Here is a very simple example, assuming we already have some data `x` and `y`: ::
+The package requires some x and y data to fit. You need to specify either a) some initial breakpoint guesses as `start_values` or b) how many breakpoints you want to fit as `n_breakpoints` (or both). Here is an elementary example, assuming we already have some data `x` and `y`: ::
 
 	import piecewise_regression
 	pw_fit = piecewise_regression.Fit(x, y, n_breakpoints=2)
@@ -57,8 +57,9 @@ The package requires some x and y data to fit. You also need to specify either a
 
 Example
 ========================
+*For demonstration purposes, substitute with your own data to fit.*
 
-Here is a more detailed example. We start off generating some data with a breakpoint. This is for demonstration purposes, normally you will have your own data to fit: ::
+1. Start-off generating some data with a breakpoint: ::
 
 	import piecewise_regression
 	import numpy as np
@@ -73,7 +74,7 @@ Here is a more detailed example. We start off generating some data with a breakp
 	yy = constant + alpha_1*xx + (alpha_2-alpha_1) * np.maximum(xx - breakpoint_1, 0) + np.random.normal(size=n_points)
 
 
-Now we fit the model: ::
+2. Fit the model: ::
 
     # Given some data, fit the model
     pw_fit = piecewise_regression.Fit(xx, yy, start_values=[5], n_breakpoints=1)
@@ -111,7 +112,7 @@ Example output: ::
 
 This includes estimates for all the model variables, along with confidence intervals. The Davies test is a hypothesis test for the existence of at least one breakpoint, against the null hypothesis of no breakpoints.  
 
-There are also tools for plotting data: ::
+3. Optional: Plotting the data and model results: ::
 
 	import matplotlib.pyplot as plt
 
@@ -140,20 +141,23 @@ You can extract data as well: ::
 How It Works
 ======================
 
-The package implements Muggeo's iterative algorithm (Muggeo "Estimating regression models with unknown break-points" (2003)), to quickly find breakpoints. That method simultaneously fits breakpoint positions and the linear models for the different segments of the fit. This method is quick and it gives confidence intervals for all the model estimates. See the accompanying paper for more details.
[…]
