### 2021-11-22 #3859 @galessiorob
Thank you @vyasr @Ebedthan @htjb for your thorough and kind reviews! We're getting very close to the finish line for this paper 🚀

@chasmani Could you please take a look at @htjb 's suggestions on their last issue, please? I don't think the PEP8 compatibility issue should be a blocker for publishing but will check with other editors. As for the formatting it'd be great if you could get to it 🎉

I created a branch where I am adding some edits, I'll open a PR once I'm done. Thanks for all the work thus far, this is such a good and useful piece of software.
