+++ piecewise_regression/main.py
@@ -15,7 +15,7 @@ try:
         validate_positive_integer,
         validate_list_of_numbers,
         validate_non_negative_integer
-        )
+    )
 except ImportError:
     import davies
     import r_squared_calc
@@ -25,7 +25,7 @@ except ImportError:
         validate_positive_integer,
         validate_list_of_numbers,
         validate_non_negative_integer
-        )
+    )
 
 
 class NextBreakpoints:
@@ -114,11 +114,12 @@ class NextBreakpoints:
 
         # First two params are a and c in the line equation
         # Beta hats are the next group of params, same length as breakpoints
-        beta_hats = results.params[2:2+len(self.current_breakpoints)]
+        beta_hats = results.params[2:2 + len(self.current_breakpoints)]
         # Gamma hats are the last group of params, same length as breakpoints
-        gamma_hats = results.params[2+len(self.current_breakpoints):]
+        gamma_hats = results.params[2 + len(self.current_breakpoints):]
         # The next breakpoints are calculated iteratively
-        self.next_breakpoints = self.current_breakpoints - gamma_hats/beta_hats
+        self.next_breakpoints = self.current_breakpoints \
+            - gamma_hats / beta_hats
 
     def calculate_all_estimates(self):
         """
@@ -128,20 +129,20 @@ class NextBreakpoints:
 
         # Extract the exstimates form the correct locations in the params
         const_estimate = params[0]
-        beta_estimates = params[2:self.n_breakpoints+2]
+        beta_estimates = params[2:self.n_breakpoints + 2]
         breakpoint_estimates = self.next_breakpoints
 
         self.estimates["const"] = {"estimate": const_estimate}
         for bp_i in range(self.n_breakpoints):
             self.estimates["beta{}".format(
-                bp_i+1)] = {"estimate": beta_estimates[bp_i]}
+                bp_i + 1)] = {"estimate": beta_estimates[bp_i]}
             self.estimates["breakpoint{}".format(
-                bp_i+1)] = {"estimate": breakpoint_estimates[bp_i]}
+                bp_i + 1)] = {"estimate": breakpoint_estimates[bp_i]}
         # Also calculate alphas
         for alpha_i in range(self.n_breakpoints + 1):
-            alpha_estimate = np.sum(params[1:alpha_i+2])
+            alpha_estimate = np.sum(params[1:alpha_i + 2])
             self.estimates["alpha{}".format(
-                alpha_i+1)] = {"estimate": alpha_estimate}
+                alpha_i + 1)] = {"estimate": alpha_estimate}
 
     def get_alpha_standard_errors(self):
         """
@@ -158,7 +159,7 @@ class NextBreakpoints:
         # var(alpha_k) = sum_{i, j} cov(alpha and betas)
         alpha_vars = []
         for alpha_n in range(self.n_breakpoints + 1):
-            alpha_cov_matrix = cov_matrix[1:alpha_n+2, 1:alpha_n+2]
+            alpha_cov_matrix = cov_matrix[1:alpha_n + 2, 1:alpha_n + 2]
             alpha_vars.append(np.sum(alpha_cov_matrix))
 
         alpha_ses = np.sqrt(alpha_vars)
@@ -192,8 +193,8 @@ class NextBreakpoints:
             # opposite to Muggeo, this is because the gamma is defined with
             # the opposite sign
             # The calculation is equivalent to Muggeos.
-            bp_var = (gamma_var + beta_var * (gamma/beta)**2 -
-                      2*(gamma/beta)*gamma_beta_covar)/(beta**2)
+            bp_var = (gamma_var + beta_var * (gamma / beta)**2 -
+                      2 * (gamma / beta) * gamma_beta_covar) / (beta**2)
             bp_vars.append(bp_var)
 
         bp_ses = np.sqrt(bp_vars)
[…]
