+++ setup_notes.txt
@@ -1,7 +1,7 @@
 To send a new version to the package manager:
 
 1. Make sure everything works
-2. Update the version number in setup.py
+2. Update the version number in setup.py and README.rst
 3. 'python setup.py sdist bdist_wheel' - this creates a dist package in the dist/ folder
 4. 'twine upload --skip-existing dist/*'
 
