### 2021-11-23 #3859 @chasmani
@hjtb thanks for looking at it again. I've just made it completely PEP8 compliant, I think. Running `flake8 piecewise_regression/ --select=E,F` now returns nothing. I'm new to PEP8 but yes I agree its a good standard to have, and I'll be using it in the future.

Thank you @galessiorob, @hjtb @vyasr and @Ebedthan for all your comments and reviews.

I think that is everything now :)
