### 2021-11-23 #3859 @vyasr
I just went through and double-checked that all my suggestions have been implemented. I've checked all the boxes and I don't have anything further to add. Both the repo and the paper are looking good!
