### 2021-11-24 #3859 @htjb
> @hjtb thanks for looking at it again. I've just made it completely PEP8 compliant, I think. Running `flake8 piecewise_regression/ --select=E,F` now returns nothing. I'm new to PEP8 but yes I agree its a good standard to have, and I'll be using it in the future.

Brilliant thank you! This looks great. Congrats on a nice piece of software :smile:
