### 2021-11-29 #3859 @galessiorob
@chasmani I put together a set of edits for the paper and the readme in [this PR.](https://github.com/chasmani/piecewise-regression/pull/2) Once you've reviewed them please deposit your software in Zenodo and post the DOI and most recent version here. After that, we can set the final version, generate a final proof, and have an Editor in Chief do the publishing.

@vyasr, @htjb, @Ebedthan Thank you so much for lending your expertise and reviewing this paper 🚀
