### 2021-11-30 #3859 @chasmani
Thanks @galessiorob for your input. I merged most of your suggestions. Thank you especially for the contribution guidelines. And thank you again to the reviewers.

I just depositied the software on zenodo and made a new release on Github.

- Zenodo DI is DOI: 10.5281/zenodo.5742317
https://zenodo.org/account/settings/github/repository/chasmani/piecewise-regression
- Version number is 1.2.0.
https://github.com/chasmani/piecewise-regression/releases/tag/v1.2.0
https://pypi.org/project/piecewise-regression/

All good?
