### 2021-12-01 #3859 @chasmani
@galessiorob I'd prefer to keep it lowercase. "piecewise-regression" is the python package name so I wanted to use the python convention of lowercase.

This is common in JOSS, see:
https://joss.theoj.org/papers/10.21105/joss.03619
https://joss.theoj.org/papers/10.21105/joss.03781
https://joss.theoj.org/papers/10.21105/joss.03358
