### 2021-12-01 #3859 @kthyng
@chasmani Hi, I'm taking over from here. Can you update the metadata in your Zenodo archive so that the title and author list exactly match your JOSS paper?
