### 2021-12-01 #3859 @kthyng
@chasmani Regarding your paper I have a few comments:
* Unless I missed it, you don't present your software package by name on the first page of your paper. Please modify to do this in the Summary and probably also Statement of Need.
* Fix capitalization in your references by adding {} around characters in your .bib file. For example, R and Python aren't always capitalized, but please look through carefully and correct all errors.
