### 2021-12-02 e9b9d72
Added referecne to the package on first page and fixed formatting in refrences
