+++ paper/paper.bib
@@ -10,7 +10,7 @@
 }
 
 @article{muggeo2008segmented,
-  title={Segmented: an R package to fit regression models with broken-line relationships},
+  title={Segmented: an {R} package to fit regression models with broken-line relationships},
   author={Muggeo, Vito MR and others},
   journal={R news},
   volume={8},
@@ -42,7 +42,7 @@
 }
 
 @article{virtanen2020scipy,
-  title={SciPy 1.0: fundamental algorithms for scientific computing in Python},
+  title={SciPy 1.0: fundamental algorithms for scientific computing in {Python}},
   author={Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and others},
   journal={Nature methods},
   volume={17},
@@ -53,7 +53,7 @@
 }
 
 @article{newville2016lmfit,
-  title={LMFIT: Non-linear least-square minimization and curve-fitting for Python},
+  title={LMFIT: Non-linear least-square minimization and curve-fitting for {Python}},
   author={Newville, Matthew and Stensitzki, Till and Allen, Daniel B and Rawlik, Michal and Ingargiola, Antonino and Nelson, Andrew},
   journal={Astrophysics Source Code Library},
   pages={ascl--1606},
@@ -61,7 +61,7 @@
 }
 
 @article{jekel2019pwlf,
-  title={PWLF: a python library for fitting 1D continuous piecewise linear functions},
+  title={PWLF: a {Python} library for fitting 1D continuous piecewise linear functions},
   author={Jekel, Charles F and Venter, Gerhard},
   journal={URL: https://github. com/cjekel/piecewise\_linear\_fit\_py},
   year={2019}
@@ -82,7 +82,7 @@
 
 
 @inproceedings{seabold2010statsmodels,
-  title={Statsmodels: Econometric and statistical modeling with python},
+  title={Statsmodels: Econometric and statistical modeling with {Python}},
   author={Seabold, Skipper and Perktold, Josef},
   booktitle={Proceedings of the 9th Python in Science Conference},
   volume={57},
@@ -124,9 +124,3 @@
   publisher={Wiley Online Library}
 }
 
-@article{jekel2019pwlf,
-  title={PWLF: a python library for fitting 1D continuous piecewise linear functions},
-  author={Jekel, Charles F and Venter, Gerhard},
-  journal={URL: https://github. com/cjekel/piecewise\_linear\_fit\_py},
-  year={2019}
-}
