### 2021-12-02 #3859 @chasmani
@kthyng thank you for taking a look at this.

I've completed the changes:
- Can you update the metadata in your Zenodo archive so that the title and author list exactly match your JOSS paper?
* DONE. The title and author names match exactly now. *
* In the paper I have two author affiliations. In zenodo I only included my primary affilation, because there was only room for one.*
- you don't present your software package by name on the first page of your paper. Please modify to do this in the Summary and probably also Statement of Need
* DONE *
- Fix capitalization in your references by adding {} around characters in your .bib file. For example, R and Python aren't always capitalized, but please look through carefully and correct all errors.
* DONE *

Hopefully that is all good. Let me know if anything else needs changing.
