### 2021-12-02 #3859 @chasmani
Many thanks to JOSS and editors @kthyng, @galessiorob and reviewers @vyasr, @htjb, and @Ebedthan - you all did a great job I thought and really helped improve the paper and package.
